feat: add Tailwind Preflight base styles
Among available reset/normalize CSS solutions, it is decided to choose
Tailwind Preflight set of base styles, because the author expects to
use the Tailwind CSS Framework for future projects.

Since The Odin Project curriculum discourages the use of any frameworks
in early projects, it is decided to include Preflight style sheet only.